Refactoring package to put dataclass objects in models/ directory
