+++ src/anaplan_api/ImportTask.py
@@ -6,7 +6,7 @@ from .ParameterAction import ParameterAction
 from .ImportParser import ImportParser
 
 if TYPE_CHECKING:
-    from .AnaplanConnection import AnaplanConnection
+    from src.anaplan_api.models.AnaplanConnection import AnaplanConnection
     from .Parser import Parser
