+++ src/anaplan_api/models/ModelDetails.py
