+++ src/anaplan_api/models/WorkspaceDetails.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from .util.strtobool import strtobool
+from src.anaplan_api.util.strtobool import strtobool
 
 
 @dataclass
